+++ src/generators/sqlitegenerator.cpp
@@ -19,8 +19,8 @@
 **************************************************************************/
 
 #include "sqlitegenerator.h"
-#include "table.h"
-#include "tablemodel.h"
+#include "../table.h"
+#include "../tablemodel.h"
 
 NUT_BEGIN_NAMESPACE
