+++ src/generators/sqlitegenerator.h
